Refactor: Post 용 DTO로 이름 변경
